refactor(lsp): split fourslash parsing helpers
